Huge refactoring of project structure
