+++ src/app/models/exceptions.py
